+++ tests/DiffTest.php
@@ -23,6 +23,10 @@ final class DiffTest extends TestCase
         $to   = 'line2a';
         $diff = new Diff($from, $to);
 
+        $this->assertSame($from, $diff->from());
+        $this->assertSame($to, $diff->to());
+        $this->assertSame([], $diff->chunks(), 'Expect chunks to be default value "array()".');
+
         $this->assertSame($from, $diff->getFrom());
         $this->assertSame($to, $diff->getTo());
         $this->assertSame([], $diff->getChunks(), 'Expect chunks to be default value "array()".');
@@ -36,6 +40,10 @@ final class DiffTest extends TestCase
 
         $diff = new Diff($from, $to, $chunks);
 
+        $this->assertSame($from, $diff->from());
+        $this->assertSame($to, $diff->to());
+        $this->assertSame($chunks, $diff->chunks(), 'Expect chunks to be passed value.');
+
         $this->assertSame($from, $diff->getFrom());
         $this->assertSame($to, $diff->getTo());
         $this->assertSame($chunks, $diff->getChunks(), 'Expect chunks to be passed value.');
@@ -44,10 +52,14 @@ final class DiffTest extends TestCase
     public function testSetChunksAfterConstruction(): void
     {
         $diff = new Diff('line1c', 'line2c');
+
+        $this->assertSame([], $diff->chunks(), 'Expect chunks to be default value "array()".');
         $this->assertSame([], $diff->getChunks(), 'Expect chunks to be default value "array()".');
 
         $chunks = [new Chunk, new Chunk(2, 3)];
         $diff->setChunks($chunks);
+
+        $this->assertSame($chunks, $diff->chunks(), 'Expect chunks to be passed value.');
         $this->assertSame($chunks, $diff->getChunks(), 'Expect chunks to be passed value.');
     }
 }